LayoutTests/http/tests/websocket/tests/hixie76/frame-length-longer-than-buffer.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Make sure WebSocket correctly skip lengthed frame even if received data has incomplete frame.");

window.jsTestIsAsync = true;

var received_messages = [];
var expected_messages = ["hello"];
function finish() {
    clearTimeout(timeoutID);
    debug(received_messages.length);
    for (var i = 0; i < received_messages; i++) {
        debug("received[" + i + "]=" + received_messages[i]);
    }

    shouldBeTrue("areArraysEqual(received_messages, expected_messages)");

    finishJSTest();
}

var ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/hixie76/frame-length-longer-than-buffer");
ws.onopen = function () {
    debug("WebSocket is open");
};
ws.onmessage = function (evt) {
    debug("received:" + evt.data);
    received_messages.push(evt.data);
};
ws.onclose = function () {
    debug("WebSocket is closed");
    finish();
};
var timeoutID = setTimeout("finish()", 2000);

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
